suggest: view model
